Moved Files to Root Directory per Instructor Critique
