--- KnightRanker_RESTful_Prototype/src/main/java/edu/calvin/cs262/pilot/knight_ranker/Token.java
@@ -1,21 +1,0 @@
-package edu.calvin.cs262.pilot.knight_ranker;
-
-public class Token {
-    private String token;
-
-    public Token() {
-
-    }
-
-    public Token(String token) {
-        this.token = token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return token;
-    }
-}
