+++ src/main/java/edu/calvin/cs262/pilot/knight_ranker/Follow.java
