+++ src/main/java/edu/calvin/cs262/pilot/knight_ranker/SecureTokenGenerator.java
@@ -1,0 +1,27 @@
+package edu.calvin.cs262.pilot.knight_ranker;
+
+import java.security.SecureRandom;
+
+// https://stackoverflow.com/questions/13992972/how-to-create-a-authentication-token-using-java
+public class SecureTokenGenerator {
+    private static final String CHARACTERS = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+
+    // 2048 bit keys should be secure until 2030 - https://web.archive.org/web/20170417095741/https://www.emc.com/emc-plus/rsa-labs/historical/twirl-and-rsa-key-size.htm
+    private static final int SECURE_TOKEN_LENGTH = 256;
+
+    private static final SecureRandom random = new SecureRandom();
+
+    private static final char[] symbols = CHARACTERS.toCharArray();
+
+    private static final char[] buf = new char[SECURE_TOKEN_LENGTH];
+
+    /**
+     * Generate the next secure random token in the series.
+     * @return String
+     */
+    public static String nextToken() {
+        for (int idx = 0; idx < buf.length; ++idx)
+            buf[idx] = symbols[random.nextInt(symbols.length)];
+        return new String(buf);
+    }
+}
