+++ src/main/java/edu/calvin/cs262/pilot/knight_ranker/Token.java
@@ -1,0 +1,21 @@
+package edu.calvin.cs262.pilot.knight_ranker;
+
+public class Token {
+    private String token;
+
+    public Token() {
+
+    }
+
+    public Token(String token) {
+        this.token = token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public String getToken() {
+        return token;
+    }
+}
